2D game (2024)
This was made during 2024, a project that was given to make a game from scratch. Although this was my first time making a game in 2D, I actually am focusing more on 3D and majority of my projects are still on pause since it requires me to really put my time on it.